fix screen transition rectangle size (#97)
Use the window size instead of the device size when drawing the transition animation over the screen.

Often there is a mismatch between the size of the device and the size of the window we are drawing to in open GL. When there is a mismatch, the fade out drawn during the transition looks off. Now we always use the window size when drawing the transition.