Don't Collect Resources When Switching Screens (#99)
Split the operation used to update resource reference counts from the operation used to clean up unneeded references.

Whenever a player dies we reset the screen with a new copy of the game screen. This results in all of the resources being returned to the resource manager. When the replacement screen is pushed onto the screen stack, all of the resources need to be reloaded from disk. This causes noticeable jank on mobile devices. By deferring the cleanup to after we request resources for the new screen, we avoid reloading resources that are used by both screens.